fix(client): WIP make pdf export work again
The PDF export feature didn't work because we were using a library that we don't have the license to (license cost $1500 per dev after update in summer 2023). This is a workaround using different libraries to make the pdf export functional again. This is WIP because there are still a number of problems with it; the content is duplictaed on screen, and it does not paginate.

Issue: https://github.sfu.ca/bfraser/415-HHA-Haiti/issues/711