chore(frontend): remove ui for pdf export
As decided in the previous meeting, the pdf export functionality does not do enough to justify maintaining it. So the button for it has been removed. Some of the underlying code has been left in, just in case it is useful for exports of the analysis feature.

Issue: https://github.sfu.ca/bfraser/415-HHA-Haiti/issues/711