Fix(test): Change case study test to be more consistent.
Before it got the first case study it got from getting all reports and checked if that was the one newly added, which does not behave consistently on the CI/CD. For now it will do this from a blank DB so that the first case study has to be the newly created one. A better test will be designed in the future.

Issue: https://github.sfu.ca/bfraser/415-HHA-Haiti/issues/714