Refactor(test): Update case study tests to use new test system.
There were some inconsistencies in the results for the testing because it was using the index of the result to check for results. This has been changed to querying the DB directly after the post has been made to see if it has been saved.

Issue: https://github.sfu.ca/bfraser/415-HHA-Haiti/issues/714