feat(analytics): wip persist chart state for session
This is a WIP implementation of using session storage to save the analytics state while the user refreshes or clicks away. However, it does not work for the question selection yet, since I get a "no department id found" error when I try to store it with the session. It works for the other 3 dropdowns.

Issue: https://github.sfu.ca/bfraser/415-HHA-Haiti/issues/763