feat(analytics): implement chart title
For the purpose of exports, there should be a quick summary of what the chart is doing. This will also help someone quickly recreate the chart by showing the configuration. There is a default suggested title, but if the user manually overrides the title, it should not change itself.

Issue: https://github.sfu.ca/bfraser/415-HHA-Haiti/issues/758